Move fixtures to root conftest
