+++ tests/python/conftest.py
@@ -136,6 +136,22 @@ def wrapper_cdm600(default_main_config: MainConfig) -> Wrapper:
     return wrapper
 
 
+@pytest.mark.usefixtures("test_db")
+@pytest.fixture(scope='function')
+def cdm531_wrapper_with_tables_created(wrapper_cdm531: Wrapper) -> Wrapper:
+    wrapper_cdm531.create_schemas()
+    wrapper_cdm531.create_cdm()
+    return wrapper_cdm531
+
+
+@pytest.mark.usefixtures("test_db")
+@pytest.fixture(scope='function')
+def cdm600_wrapper_with_tables_created(wrapper_cdm600: Wrapper) -> Wrapper:
+    wrapper_cdm600.create_schemas()
+    wrapper_cdm600.create_cdm()
+    return wrapper_cdm600
+
+
 @pytest.fixture(scope="session")
 def source_config_dir(test_data_dir: Path) -> Path:
     return test_data_dir / 'source_data_configs'
